refactor: move newtype definitions to sys module.
* re-export and document them from main crate.
